requirements.txt
asgiref==3.3.1
Django==3.1.4
mysqlclient==2.0.2
Pillow==8.0.1
PyMySQL==0.10.1
pytz==2020.4
sqlparse==0.4.1
gunicorn==19.6.0